+++ knf/knf.go
@@ -1,4 +1,4 @@
-// Package knf provides methods for working with configs in KNF format
+// Package knf provides methods for working with configuration files in KNF format
 package knf
 
 // ////////////////////////////////////////////////////////////////////////////////// //
@@ -9,28 +9,13 @@ package knf
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 import (
-	"bufio"
 	"errors"
-	"io"
 	"os"
-	"path"
-	"regexp"
 	"strconv"
 	"strings"
+	"sync"
 	"time"
 )
-
-// ////////////////////////////////////////////////////////////////////////////////// //
-
-const (
-	_COMMENT_SYMBOL   = "#"
-	_SECTION_SYMBOL   = "["
-	_SEPARATOR_SYMBOL = ":"
-	_MACRO_SYMBOL     = "{"
-	_DELIMITER        = ":"
-)
-
-const _MACRO_REGEXP = "{([a-zA-Z0-9_-]{2,}):([a-zA-Z0-9_-]{2,})}"
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
@@ -40,6 +25,8 @@ type Config struct {
 	props    []string
 	data     map[string]string
 	file     string
+
+	mx *sync.RWMutex
 }
 
 // Validator is config property validator struct
@@ -54,16 +41,20 @@ type PropertyValidator func(config *Config, prop string, value interface{}) erro
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// RegExp struct for searching and parsing macroses
-var macroRE = regexp.MustCompile(_MACRO_REGEXP)
+var (
+	ErrConfigIsNil = errors.New("Config struct is nil")
+	ErrFileNotSet  = errors.New("Path to config file is empty (non initialized struct?)")
+)
 
-// Global config struct
+// ////////////////////////////////////////////////////////////////////////////////// //
+
+// global is global configuration file
 var global *Config
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// Global reads and parses config file
-// Global config will be accessible globally from any part of the code
+// Global reads and parses configuration file
+// Global instance is accessible from any part of the code
 func Global(file string) error {
 	config, err := Read(file)
 
@@ -76,40 +67,21 @@ func Global(file string) error {
 	return nil
 }
 
-// Read reads and parse config file
+// Read reads and parses configuration file
 func Read(file string) (*Config, error) {
-	fd, err := os.OpenFile(path.Clean(file), os.O_RDONLY, 0)
-
-	if err != nil {
-		return nil, err
-	}
-
-	defer fd.Close()
-
-	config := &Config{
-		data: make(map[string]string),
-		file: file,
-	}
-
-	err = readConfigData(config, fd, file)
-
-	if err != nil {
-		return nil, err
-	}
-
-	return config, nil
+	return readKNFFile(file)
 }
 
-// Reload reads and parse global config file
+// Reload reloads global configuration file
 func Reload() (map[string]bool, error) {
 	if global == nil {
-		return nil, errors.New("Global config is not loaded")
+		return nil, ErrConfigIsNil
 	}
 
 	return global.Reload()
 }
 
-// GetS returns global config value as string
+// GetS returns configuration value as string
 func GetS(name string, defvals ...string) string {
 	if global == nil {
 		if len(defvals) == 0 {
@@ -122,7 +94,7 @@ func GetS(name string, defvals ...string) string {
 	return global.GetS(name, defvals...)
 }
 
-// GetI returns global config value as int
+// GetI returns configuration value as int
 func GetI(name string, defvals ...int) int {
 	if global == nil {
 		if len(defvals) == 0 {
@@ -135,7 +107,7 @@ func GetI(name string, defvals ...int) int {
 	return global.GetI(name, defvals...)
 }
 
-// GetI64 returns global config value as int64
+// GetI64 returns configuration value as int64
 func GetI64(name string, defvals ...int64) int64 {
 	if global == nil {
 		if len(defvals) == 0 {
@@ -148,7 +120,7 @@ func GetI64(name string, defvals ...int64) int64 {
 	return global.GetI64(name, defvals...)
 }
 
-// GetU returns global config value as uint
+// GetU returns configuration value as uint
 func GetU(name string, defvals ...uint) uint {
 	if global == nil {
 		if len(defvals) == 0 {
@@ -161,7 +133,7 @@ func GetU(name string, defvals ...uint) uint {
 	return global.GetU(name, defvals...)
 }
 
-// GetU64 returns global config value as uint64
+// GetU64 returns configuration value as uint64
 func GetU64(name string, defvals ...uint64) uint64 {
 	if global == nil {
 		if len(defvals) == 0 {
@@ -174,7 +146,7 @@ func GetU64(name string, defvals ...uint64) uint64 {
 	return global.GetU64(name, defvals...)
 }
 
-// GetF returns global config value as floating number
+// GetF returns configuration value as floating number
 func GetF(name string, defvals ...float64) float64 {
 	if global == nil {
 		if len(defvals) == 0 {
@@ -187,7 +159,7 @@ func GetF(name string, defvals ...float64) float64 {
 	return global.GetF(name, defvals...)
 }
 
-// GetB returns global config value as boolean
+// GetB returns configuration value as boolean
 func GetB(name string, defvals ...bool) bool {
 	if global == nil {
 		if len(defvals) == 0 {
@@ -200,7 +172,7 @@ func GetB(name string, defvals ...bool) bool {
 	return global.GetB(name, defvals...)
 }
 
-// GetM returns global config value as file mode
+// GetM returns configuration value as file mode
 func GetM(name string, defvals ...os.FileMode) os.FileMode {
 	if global == nil {
 		if len(defvals) == 0 {
@@ -213,7 +185,7 @@ func GetM(name string, defvals ...os.FileMode) os.FileMode {
 	return global.GetM(name, defvals...)
 }
 
-// GetD returns global config values as duration
+// GetD returns configuration values as duration
 func GetD(name string, defvals ...time.Duration) time.Duration {
 	if global == nil {
 		if len(defvals) == 0 {
@@ -226,7 +198,7 @@ func GetD(name string, defvals ...time.Duration) time.Duration {
 	return global.GetD(name, defvals...)
 }
 
-// HasSection check if section exist
+// HasSection checks if section exist
 func HasSection(section string) bool {
 	if global == nil {
 		return false
@@ -235,7 +207,7 @@ func HasSection(section string) bool {
 	return global.HasSection(section)
 }
 
-// HasProp check if property exist
+// HasProp checks if property exist
 func HasProp(name string) bool {
 	if global == nil {
 		return false
@@ -247,7 +219,7 @@ func HasProp(name string) bool {
 // Sections returns slice with section names
 func Sections() []string {
 	if global == nil {
-		return []string{}
+		return nil
 	}
 
 	return global.Sections()
@@ -256,17 +228,17 @@ func Sections() []string {
 // Props returns slice with properties names in some section
 func Props(section string) []string {
 	if global == nil {
-		return []string{}
+		return nil
 	}
 
 	return global.Props(section)
 }
 
-// Validate require slice with pointers to validators and
+// Validate executes all given validators and
 // returns slice with validation errors
 func Validate(validators []*Validator) []error {
 	if global == nil {
-		return []error{errors.New("Global config struct is nil")}
+		return []error{ErrConfigIsNil}
 	}
 
 	return global.Validate(validators)
@@ -274,14 +246,14 @@ func Validate(validators []*Validator) []error {
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-// Reload read and parse config file
+// Reload reloads configuration file
 func (c *Config) Reload() (map[string]bool, error) {
-	if c == nil {
-		return nil, errors.New("Config is nil")
+	if c == nil || c.mx == nil {
+		return nil, ErrConfigIsNil
 	}
 
 	if c.file == "" {
-		return nil, errors.New("Path to config file is empty (non initialized struct?)")
+		return nil, ErrFileNotSet
 	}
 
 	nc, err := Read(c.file)
@@ -292,18 +264,25 @@ func (c *Config) Reload() (map[string]bool, error) {
 
 	changes := make(map[string]bool)
 
-	for prop, value := range c.data {
-		changes[prop] = value != nc.data[prop]
+	c.mx.RLock()
+
+	for _, prop := range c.props {
+		changes[prop] = c.GetS(prop) != nc.GetS(prop)
 	}
 
-	c.data, c.sections = nc.data, nc.sections
+	c.mx.RUnlock()
+	c.mx.Lock()
 
+	// Update current config data
+	c.data, c.sections, c.props = nc.data, nc.sections, nc.props
+
+	c.mx.Unlock()
 	return changes, nil
 }
 
-// GetS returns config value as string
+// GetS returns configuration value as string
 func (c *Config) GetS(name string, defvals ...string) string {
-	if c == nil {
+	if c == nil || c.mx == nil {
 		if len(defvals) == 0 {
 			return ""
 		}
@@ -311,7 +290,9 @@ func (c *Config) GetS(name string, defvals ...string) string {
 		return defvals[0]
 	}
 
-	val := c.data[name]
+	c.mx.RLock()
+	val := c.data[strings.ToLower(name)]
+	c.mx.RUnlock()
 
 	if val == "" {
 		if len(defvals) == 0 {
@@ -324,9 +305,9 @@ func (c *Config) GetS(name string, defvals ...string) string {
 	return val
 }
 
-// GetI64 returns config value as int64
+// GetI64 returns configuration value as int64
 func (c *Config) GetI64(name string, defvals ...int64) int64 {
-	if c == nil {
+	if c == nil || c.mx == nil {
 		if len(defvals) == 0 {
 			return 0
 		}
@@ -334,7 +315,9 @@ func (c *Config) GetI64(name string, defvals ...int64) int64 {
 		return defvals[0]
 	}
 
-	val := c.data[name]
+	c.mx.RLock()
+	val := c.data[strings.ToLower(name)]
+	c.mx.RUnlock()
 
 	if val == "" {
 		if len(defvals) == 0 {
@@ -364,7 +347,7 @@ func (c *Config) GetI64(name string, defvals ...int64) int64 {
 	return valInt
 }
 
-// GetI returns config value as int
+// GetI returns configuration value as int
 func (c *Config) GetI(name string, defvals ...int) int {
 	if len(defvals) != 0 {
 		return int(c.GetI64(name, int64(defvals[0])))
@@ -373,7 +356,7 @@ func (c *Config) GetI(name string, defvals ...int) int {
 	return int(c.GetI64(name))
 }
 
-// GetU returns config value as uint
+// GetU returns configuration value as uint
 func (c *Config) GetU(name string, defvals ...uint) uint {
 	if len(defvals) != 0 {
 		return uint(c.GetI64(name, int64(defvals[0])))
@@ -382,7 +365,7 @@ func (c *Config) GetU(name string, defvals ...uint) uint {
 	return uint(c.GetI64(name))
 }
 
-// GetU64 returns config value as uint64
+// GetU64 returns configuration value as uint64
 func (c *Config) GetU64(name string, defvals ...uint64) uint64 {
 	if len(defvals) != 0 {
 		return uint64(c.GetI64(name, int64(defvals[0])))
@@ -391,9 +374,9 @@ func (c *Config) GetU64(name string, defvals ...uint64) uint64 {
 	return uint64(c.GetI64(name))
 }
 
-// GetF returns config value as floating number
+// GetF returns configuration value as floating number
 func (c *Config) GetF(name string, defvals ...float64) float64 {
-	if c == nil {
+	if c == nil || c.mx == nil {
 		if len(defvals) == 0 {
 			return 0.0
 		}
@@ -401,7 +384,9 @@ func (c *Config) GetF(name string, defvals ...float64) float64 {
 		return defvals[0]
 	}
 
-	val := c.data[name]
+	c.mx.RLock()
+	val := c.data[strings.ToLower(name)]
+	c.mx.RUnlock()
 
 	if val == "" {
 		if len(defvals) == 0 {
@@ -420,9 +405,9 @@ func (c *Config) GetF(name string, defvals ...float64) float64 {
 	return valFl
 }
 
-// GetB returns config value as boolean
+// GetB returns configuration value as boolean
 func (c *Config) GetB(name string, defvals ...bool) bool {
-	if c == nil {
+	if c == nil || c.mx == nil {
 		if len(defvals) == 0 {
 			return false
 		}
@@ -430,7 +415,9 @@ func (c *Config) GetB(name string, defvals ...bool) bool {
 		return defvals[0]
 	}
 
-	val := c.data[name]
+	c.mx.RLock()
+	val := c.data[strings.ToLower(name)]
+	c.mx.RUnlock()
 
 	if val == "" {
 		if len(defvals) == 0 {
@@ -448,9 +435,9 @@ func (c *Config) GetB(name string, defvals ...bool) bool {
 	}
 }
 
-// GetM returns config value as file mode
+// GetM returns configuration value as file mode
 func (c *Config) GetM(name string, defvals ...os.FileMode) os.FileMode {
-	if c == nil {
+	if c == nil || c.mx == nil {
 		if len(defvals) == 0 {
 			return os.FileMode(0)
 		}
@@ -458,7 +445,9 @@ func (c *Config) GetM(name string, defvals ...os.FileMode) os.FileMode {
 		return defvals[0]
 	}
 
-	val := c.data[name]
+	c.mx.RLock()
+	val := c.data[strings.ToLower(name)]
+	c.mx.RUnlock()
 
 	if val == "" {
 		if len(defvals) == 0 {
@@ -477,9 +466,9 @@ func (c *Config) GetM(name string, defvals ...os.FileMode) os.FileMode {
 	return os.FileMode(valM)
 }
 
-// GetD returns config value as duration
+// GetD returns configuration value as duration
 func (c *Config) GetD(name string, defvals ...time.Duration) time.Duration {
-	if c == nil {
+	if c == nil || c.mx == nil {
 		if len(defvals) == 0 {
 			return time.Duration(0)
 		}
@@ -487,7 +476,9 @@ func (c *Config) GetD(name string, defvals ...time.Duration) time.Duration {
 		return defvals[0]
 	}
 
-	val := c.data[name]
+	c.mx.RLock()
+	val := c.data[strings.ToLower(name)]
+	c.mx.RUnlock()
 
 	if val == "" {
 		if len(defvals) == 0 {
@@ -500,53 +491,66 @@ func (c *Config) GetD(name string, defvals ...time.Duration) time.Duration {
 	return time.Duration(c.GetI64(name)) * time.Second
 }
 
-// HasSection check if section exist
+// HasSection checks if section exist
 func (c *Config) HasSection(section string) bool {
-	if c == nil {
+	if c == nil || c.mx == nil {
 		return false
 	}
 
-	return c.data[section+_DELIMITER] == "true"
+	c.mx.RLock()
+	defer c.mx.RUnlock()
+
+	return c.data[strings.ToLower(section)] == "!"
 }
 
-// HasProp check if property exist
+// HasProp checks if property exist
 func (c *Config) HasProp(name string) bool {
-	if c == nil {
+	if c == nil || c.mx == nil {
 		return false
 	}
 
-	return c.data[name] != ""
+	c.mx.RLock()
+	defer c.mx.RUnlock()
+
+	return c.data[strings.ToLower(name)] != ""
 }
 
 // Sections returns slice with section names
 func (c *Config) Sections() []string {
-	if c == nil {
-		return []string{}
+	if c == nil || c.mx == nil {
+		return nil
 	}
+
+	c.mx.RLock()
+	defer c.mx.RUnlock()
 
 	return c.sections
 }
 
 // Props returns slice with properties names in some section
 func (c *Config) Props(section string) []string {
-	if c == nil || !c.HasSection(section) {
-		return []string{}
-	}
-
 	var result []string
+
+	if c == nil || !c.HasSection(section) {
+		return result
+	}
 
 	// Section name + delimiter
 	snLength := len(section) + 1
+
+	c.mx.RLock()
 
 	for _, prop := range c.props {
 		if len(prop) <= snLength {
 			continue
 		}
 
-		if prop[0:snLength] == section+_DELIMITER {
+		if prop[:snLength] == section+_PROP_DELIMITER {
 			result = append(result, prop[snLength:])
 		}
 	}
+
+	defer c.mx.RUnlock()
 
 	return result
 }
@@ -554,11 +558,13 @@ func (c *Config) Props(section string) []string {
 // Validate executes all given validators and
 // returns slice with validation errors
 func (c *Config) Validate(validators []*Validator) []error {
-	if c == nil {
-		return []error{errors.New("Config is nil")}
+	if c == nil || c.mx == nil {
+		return []error{ErrConfigIsNil}
 	}
 
 	var result []error
+
+	c.mx.RLock()
 
 	for _, v := range validators {
 		err := v.Func(c, v.Property, v.Value)
@@ -568,82 +574,9 @@ func (c *Config) Validate(validators []*Validator) []error {
 		}
 	}
 
+	defer c.mx.RUnlock()
+
 	return result
-}
-
-// ////////////////////////////////////////////////////////////////////////////////// //
-
-func readConfigData(config *Config, fd io.Reader, file string) error {
-	var sectionName = ""
-
-	reader := bufio.NewReader(fd)
-	scanner := bufio.NewScanner(reader)
-
-	var isDataRead bool
-
-	for scanner.Scan() {
-		line := scanner.Text()
-
-		isDataRead = true
-
-		if line == "" || strings.Trim(line, " \t") == "" {
-			continue
-		}
-
-		if strings.HasPrefix(strings.TrimLeft(line, " \t"), _COMMENT_SYMBOL) {
-			continue
-		}
-
-		if strings.HasPrefix(strings.TrimLeft(line, " \t"), _SECTION_SYMBOL) {
-			sectionName = strings.Trim(line, "[ ]")
-			config.data[sectionName+_DELIMITER] = "true"
-			config.sections = append(config.sections, sectionName)
-			continue
-		}
-
-		if sectionName == "" {
-			return errors.New("Configuration file " + file + " is malformed")
-		}
-
-		propName, propValue := parseRecord(line, config)
-		fullPropName := sectionName + _DELIMITER + propName
-
-		config.props = append(config.props, fullPropName)
-		config.data[fullPropName] = propValue
-	}
-
-	if !isDataRead {
-		return errors.New("Configuration file " + file + " is empty")
-	}
-
-	return scanner.Err()
-}
-
-func parseRecord(data string, config *Config) (string, string) {
-	va := strings.Split(data, _SEPARATOR_SYMBOL)
-
-	propName := va[0]
-	propValue := strings.Join(va[1:], _SEPARATOR_SYMBOL)
-
-	propName = strings.TrimLeft(propName, " \t")
-	propValue = strings.TrimLeft(propValue, " \t")
-	propValue = strings.TrimRight(propValue, " ")
-
-	if strings.Contains(propValue, _MACRO_SYMBOL) {
-		macroses := macroRE.FindAllStringSubmatch(propValue, -1)
-
-		for _, macros := range macroses {
-			macroFull := macros[0]
-			macroSect := macros[1]
-			macroProp := macros[2]
-
-			macroVal := config.GetS(macroSect + _DELIMITER + macroProp)
-
-			propValue = strings.Replace(propValue, macroFull, macroVal, -1)
-		}
-	}
-
-	return propName, propValue
 }
 
 // ////////////////////////////////////////////////////////////////////////////////// //